Fix not compiling code when AA is running first
It is possible that the AA processor is running first, then the
PermissionDispatcher processor. In that case, AA claims the @RuntimePermissions annotation,
so the latter processor cannot generate the permission handling helper class.

This commit removes the @RuntimePermissions handler from this plugin,
so it always can be claimed by PermissionDispatcher, regardless of processor order.

#7